requirements.txt
django>=4
django-computedfields
psycopg2
transifex-client
djangorestframework
djangorestframework-gis
pyyaml
uritemplate
pyproj
